Finished the dashboard -- needs cleaning
