+++ frontend/package.json
@@ -18,6 +18,7 @@
     "dotenv": "^16.3.1",
     "flowbite": "^2.2.1",
     "flowbite-react": "^0.7.2",
+    "primeicons": "^6.0.1",
     "primereact": "^10.3.3",
     "react": "^17.0.0 || ^18.0.0",
     "react-big-calendar": "^1.8.6",
