+++ frontend/src/screens/studentDashboard/index.css
@@ -15,6 +15,9 @@
 
   margin-top: 5rem;
 }
+.student-profile__icon {
+  margin-right: 1rem;
+}
 .student-information__profile {
   display: flex;
   flex-direction: row;
